refactor: replace useTransationState custom hook with direct zustand call
