--- src/renderer/src/hooks/transaction/useTransactionState.ts
@@ -1,78 +1,0 @@
-import { useBillingStore } from "@/store/billingStore";
-import { useLineItemsStore } from "@/store/lineItemsStore";
-import { useSearchDropdownStore } from "@/store/searchDropdownStore";
-
-const useTransactionState = () => {
-  // billing store
-  const billingId = useBillingStore((state) => state.billingId);
-  const setBillingId = useBillingStore((state) => state.setBillingId);
-  const billingType = useBillingStore((state) => state.billingType);
-  const setBillingType = useBillingStore((state) => state.setBillingType);
-  const transactionNo = useBillingStore((state) => state.transactionNo);
-  const setTransactionNo = useBillingStore((state) => state.setTransactionNo);
-  const billingDate = useBillingStore((state) => state.billingDate);
-  const setBillingDate = useBillingStore((state) => state.setBillingDate);
-  const customerId = useBillingStore((state) => state.customerId);
-  const setCustomerId = useBillingStore((state) => state.setCustomerId);
-  const customerName = useBillingStore((state) => state.customerName);
-  const setCustomerName = useBillingStore((state) => state.setCustomerName);
-  const customerContact = useBillingStore((state) => state.customerContact);
-  const setCustomerContact = useBillingStore((state) => state.setCustomerContact);
-  const isNewCustomer = useBillingStore((state) => state.isNewCustomer);
-  const setIsNewCustomer = useBillingStore((state) => state.setIsNewCustomer);
-  const status = useBillingStore((state) => state.status);
-  const setStatus = useBillingStore((state) => state.setStatus);
-  // lineItems store
-  const isCountColumnVisible = useLineItemsStore((state) => state.isCountColumnVisible);
-  const setIsCountControlsVisible = useLineItemsStore((state) => state.setIsCountControlsVisible);
-  const lineItems = useLineItemsStore((state) => state.lineItems);
-  const setLineItems = useLineItemsStore((state) => state.setLineItems);
-  const originalLineItems = useLineItemsStore((state) => state.originalLineItems);
-  const setOriginalLineItems = useLineItemsStore((state) => state.setOriginalLineItems);
-  const addEmptyLineItem = useLineItemsStore((state) => state.addEmptyLineItem);
-  const updateLineItem = useLineItemsStore((state) => state.updateLineItem);
-  const updateInternalIds = useLineItemsStore((state) => state.updateInternalIds);
-  const deleteLineItem = useLineItemsStore((state) => state.deleteLineItem);
-  const setAllChecked = useLineItemsStore((state) => state.setAllChecked);
-  // search drop down store
-  const setItemQuery = useSearchDropdownStore((state) => state.setItemQuery);
-  const setActiveRowId = useSearchDropdownStore((state) => state.setActiveRowId);
-  const setIsDropdownOpen = useSearchDropdownStore((state) => state.setIsDropdownOpen);
-
-  return {
-    billingId,
-    setBillingId,
-    billingType,
-    setBillingType,
-    transactionNo,
-    setTransactionNo,
-    billingDate,
-    setBillingDate,
-    customerId,
-    setCustomerId,
-    customerName,
-    setCustomerName,
-    customerContact,
-    setCustomerContact,
-    isNewCustomer,
-    setIsNewCustomer,
-    status,
-    setStatus,
-    isCountColumnVisible,
-    setIsCountControlsVisible,
-    lineItems,
-    setLineItems,
-    originalLineItems,
-    setOriginalLineItems,
-    addEmptyLineItem,
-    updateLineItem,
-    updateInternalIds,
-    deleteLineItem,
-    setAllChecked,
-    setItemQuery,
-    setActiveRowId,
-    setIsDropdownOpen
-  };
-};
-
-export default useTransactionState;
